+++ ydb/core/formats/arrow/accessor/abstract/ya.make
@@ -6,6 +6,7 @@ PEERDIR(
     ydb/library/conclusion
     ydb/services/metadata/abstract
     ydb/library/actors/core
+    ydb/core/formats/arrow
     ydb/core/formats/arrow/accessor/common
     ydb/core/formats/arrow/filter
     ydb/library/formats/arrow/protos
