+++ ydb/core/formats/arrow/common/ya.make
@@ -2,12 +2,13 @@ LIBRARY()
 
 PEERDIR(
     contrib/libs/apache/arrow
+    ydb/core/formats/arrow/accessor/plain
     ydb/core/formats/arrow/container
+    ydb/core/formats/arrow/splitter
     ydb/core/formats/arrow/switch
     ydb/library/actors/core
     ydb/library/conclusion
     ydb/library/formats/arrow
-    ydb/core/formats/arrow/splitter
 )
 
 SRCS(
