+++ ydb/core/formats/arrow/hash/calcer.cpp
@@ -1,6 +1,5 @@
 #include "calcer.h"
 
-#include <ydb/core/formats/arrow/arrow_helpers.h>
 #include <ydb/core/formats/arrow/switch/switch_type.h>
 
 #include <ydb/library/actors/core/log.h>
@@ -143,6 +142,47 @@ ui64 TXX64::CalcHash(const std::shared_ptr<arrow::Scalar>& scalar) {
     calcer.Start();
     AppendField(scalar, calcer);
     return calcer.Finish();
+}
+
+namespace {
+
+template <class TDataContainer>
+bool BuildHashUI64Impl(std::shared_ptr<TDataContainer>& batch, const std::vector<std::string>& fieldNames, const std::string& hashFieldName) {
+    if (fieldNames.size() == 0) {
+        return false;
+    }
+    Y_ABORT_UNLESS(!batch->GetColumnByName(hashFieldName));
+    if (fieldNames.size() == 1) {
+        auto column = batch->GetColumnByName(fieldNames.front());
+        if (!column) {
+            AFL_WARN(NKikimrServices::ARROW_HELPER)("event", "cannot_build_hash")("reason", "field_not_found")("field_name", fieldNames.front());
+            return false;
+        }
+        Y_ABORT_UNLESS(column);
+        if (column->type()->id() == arrow::Type::UINT64 || column->type()->id() == arrow::Type::UINT32 ||
+            column->type()->id() == arrow::Type::INT64 || column->type()->id() == arrow::Type::INT32) {
+            batch = TStatusValidator::GetValid(
+                batch->AddColumn(batch->num_columns(), std::make_shared<arrow::Field>(hashFieldName, column->type()), column));
+            return true;
+        }
+    }
+    std::shared_ptr<arrow::Array> hashColumn =
+        NArrow::NHash::TXX64(fieldNames, NArrow::NHash::TXX64::ENoColumnPolicy::Verify, 34323543).ExecuteToArray(batch);
+    batch = NAdapter::TDataBuilderPolicy<TDataContainer>::AddColumn(
+        batch, std::make_shared<arrow::Field>(hashFieldName, hashColumn->type()), hashColumn);
+    return true;
+}
+
+}   // namespace
+
+bool THashConstructor::BuildHashUI64(
+    std::shared_ptr<arrow::Table>& batch, const std::vector<std::string>& fieldNames, const std::string& hashFieldName) {
+    return BuildHashUI64Impl(batch, fieldNames, hashFieldName);
+}
+
+bool THashConstructor::BuildHashUI64(
+    std::shared_ptr<arrow::RecordBatch>& batch, const std::vector<std::string>& fieldNames, const std::string& hashFieldName) {
+    return BuildHashUI64Impl(batch, fieldNames, hashFieldName);
 }
 
 }   // namespace NKikimr::NArrow::NHash
