+++ ydb/core/formats/arrow/hash/ya.make
@@ -2,8 +2,8 @@ LIBRARY()
 
 PEERDIR(
     contrib/libs/apache/arrow
+    ydb/core/formats/arrow/common
     ydb/core/formats/arrow/switch
-    ydb/core/formats/arrow/reader
     ydb/library/actors/core
     ydb/library/services
     ydb/library/actors/protos
