+++ ydb/core/formats/arrow/reader/ya.make
@@ -2,9 +2,13 @@ LIBRARY()
 
 PEERDIR(
     contrib/libs/apache/arrow
+    ydb/core/formats/arrow/accessor/abstract
+    ydb/core/formats/arrow/common
+    ydb/core/formats/arrow/container
     ydb/core/formats/arrow/filter
     ydb/core/formats/arrow/switch
-    ydb/core/formats/arrow/container
+    ydb/core/scheme
+
     ydb/library/actors/core
     ydb/library/services
     ydb/library/formats/arrow
