+++ ydb/core/formats/arrow/ut/ya.make
@@ -7,7 +7,10 @@ PEERDIR(
     ydb/library/arrow_kernels
     ydb/library/formats/arrow/simple_builder
     ydb/core/formats/arrow/filter
+    ydb/core/formats/arrow/hash
+    ydb/core/formats/arrow/printer
     ydb/core/formats/arrow/program
+    ydb/core/formats/arrow/reader
     ydb/core/base
     ydb/library/formats/arrow
