+++ ydb/core/formats/arrow/ya.make
@@ -1,36 +1,16 @@
-RECURSE(
-    container
-    filter
-)
-
-RECURSE_FOR_TESTS(
-    ut
-)
-
 LIBRARY()
 
 PEERDIR(
     contrib/libs/apache/arrow
-    ydb/core/scheme
-    ydb/core/formats/arrow/accessor
-    ydb/core/formats/arrow/dictionary
-    ydb/core/formats/arrow/hash
-    ydb/core/formats/arrow/printer
-    ydb/core/formats/arrow/reader
-    ydb/core/formats/arrow/rows
-    ydb/core/formats/arrow/save_load
     ydb/core/formats/arrow/serializer
-    ydb/core/formats/arrow/splitter
-    ydb/core/formats/arrow/transformer
+    ydb/core/kqp/common/result_set_format
+    ydb/core/scheme
     ydb/library/actors/core
-    ydb/library/arrow_kernels
-    yql/essentials/types/binary_json
-    yql/essentials/types/dynumber
     ydb/library/formats/arrow
     ydb/library/services
-    yql/essentials/core/arrow_kernels/request
     yql/essentials/minikql
-    ydb/core/kqp/common/result_set_format
+    yql/essentials/types/binary_json
+    yql/essentials/types/dynumber
 )
 
 YQL_LAST_ABI_VERSION()
@@ -48,3 +28,21 @@ SRCS(
 )
 
 END()
+
+RECURSE(
+    accessor
+    container
+    dictionary
+    filter
+    hash
+    printer
+    reader
+    rows
+    save_load
+    splitter
+    transformer
+)
+
+RECURSE_FOR_TESTS(
+    ut
+)
