+++ ydb/core/tx/columnshard/common/ya.make
@@ -17,6 +17,7 @@ PEERDIR(
     ydb/library/formats/arrow/protos
     contrib/libs/apache/arrow
     ydb/core/formats/arrow
+    ydb/core/formats/arrow/rows
     ydb/core/tx/columnshard/common/protos
     ydb/core/tx/columnshard/data_sharing/protos
     ydb/core/tx/columnshard/transactions/protos
