+++ ydb/core/tx/columnshard/engines/reader/plain_reader/iterator/ya.make
@@ -15,7 +15,11 @@ SRCS(
 
 PEERDIR(
     ydb/core/formats/arrow
-    ydb/core/formats/arrow/filter
+    ydb/core/formats/arrow/accessor/plain
+    ydb/core/formats/arrow/hash
+    ydb/core/formats/arrow/program
+    ydb/core/formats/arrow/reader
+    ydb/core/formats/arrow/serializer
     ydb/core/tx/columnshard/blobs_action
     ydb/core/tx/columnshard/engines/reader/common_reader/iterator
     ydb/core/tx/conveyor/usage
